+++ src/main/java/in/vaidicjoshi/geektrust/backend/mymoney/dto/MyMoneyFundPortfolio.java
@@ -39,7 +39,7 @@ public class MyMoneyFundPortfolio implements Cloneable {
    *
    * @return total investment across all asset class
    */
-  public Integer getTotalInvestment() {
-    return funds.stream().mapToInt(FundEntity::getAmount).sum();
+  public double getTotalInvestment() {
+    return funds.stream().mapToDouble(FundEntity::getAmount).sum();
   }
 }
